requirements.txt
asgiref==3.7.2
Django==5.0.2
django-background-tasks==1.2.8
django-cors-headers==4.9.0
django-filter==25.1
django-modeltranslation==0.18.11
django-parler==2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.2
gunicorn==21.2.0
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-slugify==8.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
uritemplate==4.1.1
